+++ lib/utils.js
@@ -49,6 +49,10 @@ module.exports = {
 		return email.trim().toLowerCase();
 	},
 
+	ip_address: (req) => {
+		return req.headers['x-forwarded-for'] || req.connection.remoteAddress;
+	},
+
 	error_page: (rsp, type) => {
 		rsp.render('page', {
 			title: 'Error',
